File input area very large in Firefox
The file input area extended well beyond its confines in Firefox. This
commit limits the clickable area to the wdiget in Firefox, and to the
selector button in other browsers (which is good in IE because there is
a security constraint in IE that makes it invalid to click on anything
else to upload a file).