Update input mask documentation
The documentation for this is off a little bit - in the list
of characters in the "Format" section, the ? character is
listed as Alphanumeric, however this is incorrect. Instead,
anything proceeding it in the defined mask is deemed optional.
The asterisk * is then listed as Any character, but this is
actually the alphanumeric denotation, so I swapped them.